Completing / tidying WCF service

+++ party/dotnet/PartyService/Service.cs
@@ -48,10 +48,13 @@ namespace PartyService
 
                 if (r == null)
                 {
-                    throw (new PartyServiceException(string.Format("Party {0} not found", id), HttpStatusCode.NotFound));
+                    throw (new PartyServiceException(
+                        string.Format("Party {0} not found", id), 
+                        HttpStatusCode.NotFound));
                 }
 
                 string type = Convert.ToString(r["type"]);
+
                 switch (type)
                 {
                     case "P":
@@ -63,7 +66,8 @@ namespace PartyService
                         string name = Convert.ToString(r["name"]);
                         return new Business(id, name);
                     default:
-                        return null;
+                        throw (new PartyServiceException(
+                        string.Format("Unknown party type '{0}' retrieved", type)));
                 }
             }
             catch (Exception ex)
@@ -77,25 +81,33 @@ namespace PartyService
         [OperationContract]
         public PartyResults GetPartiesByName(string name)
         {
-            string baseUrl = WebOperationContext.Current.IncomingRequest.UriTemplateMatch.BaseUri.AbsoluteUri;
-
-            PartyResults pl = new PartyResults();
-
-            DataTable t = database.GetPartyByName(name);
-            if (t == null) { return pl; }
-
-            foreach (DataRow r in t.Rows)
+            try
             {
-                int id = Convert.ToInt32(r["id"]);
-                string type = Convert.ToString(r["type"]);
-                string partyName = Convert.ToString(r["name"]);
-                string link = string.Format("{0}party/{1}", baseUrl, id);
+                string baseUrl = WebOperationContext.Current.IncomingRequest.UriTemplateMatch.BaseUri.AbsoluteUri;
 
-                PartySummary ps = new PartySummary(id, type, partyName, link);
-                pl.PartyList.Add(ps);
+                PartyResults pl = new PartyResults();
+
+                DataTable t = database.GetPartyByName(name);
+
+                if (t == null) { return pl; }
+
+                foreach (DataRow r in t.Rows)
+                {
+                    int id = Convert.ToInt32(r["id"]);
+                    string type = Convert.ToString(r["type"]);
+                    string partyName = Convert.ToString(r["name"]);
+                    string link = string.Format("{0}party/{1}", baseUrl, id);
+
+                    PartySummary ps = new PartySummary(id, type, partyName, link);
+                    pl.PartyList.Add(ps);
+                }
+
+                return pl;
             }
-
-            return pl;
+            catch (Exception ex)
+            {
+                return (PartyResults)ErrorObject(new PartyResults(), ex);
+            }
         }
 
         [OperationContract]
@@ -119,13 +131,26 @@ namespace PartyService
         {
             try
             {
-                int id = Convert.ToInt32(idStr);
+                int id;
+
+                if (!int.TryParse(idStr, out id))
+                {
+                    throw (new PartyServiceException(
+                        string.Format("'{0}' is not a valid party identifier", idStr),
+                        HttpStatusCode.BadRequest));
+                }
+
                 if (database.GetParty(id) == null)
                 {
-                    throw (new PartyServiceException(string.Format("Person {0} not found", id), HttpStatusCode.NotFound));
+                    throw (new PartyServiceException(
+                        string.Format("Person {0} not found", id), 
+                        HttpStatusCode.NotFound));
                 }
+
                 database.UpdatePerson(id, person.FirstName, person.LastName, person.DateOfBirth);
+
                 return (Person)GetParty(idStr);
+
             }
             catch (Exception ex)
             {
@@ -167,7 +192,9 @@ namespace PartyService
                 {
                     throw (new Exception(string.Format("Business {0} not found", id)));
                 }
+
                 database.UpdateBusiness(id, business.Name);
+                
                 return (Business)GetParty(idStr);
             }
             catch (Exception ex)
@@ -181,10 +208,33 @@ namespace PartyService
         [WebInvoke(UriTemplate = "party/{idStr}/delete")]
         public Acknowledgement DeleteParty(string idStr)
         {
-            int id = Convert.ToInt32(idStr);
-            database.DeleteParty(id);
-            return new Acknowledgement(
-                string.Format("Party {0} deleted", id));
+
+            try
+            {
+                int id;
+
+                if (!int.TryParse(idStr, out id))
+                {
+                    throw (new PartyServiceException(
+                        string.Format("'{0}' is not a valid party identifier", idStr),
+                        HttpStatusCode.BadRequest));
+                }
+
+                if (database.GetParty(id) == null)
+                {
+                    throw (new Exception(string.Format("Party {0} not found", id)));
+                }
+
+                database.DeleteParty(id);
+
+                return new Acknowledgement(
+                    string.Format("Party {0} deleted", id));
+            }
+            catch (Exception ex)
+            {
+                return (Acknowledgement)ErrorObject(new Acknowledgement(), ex);
+            }
+
         }
 
         [WebGet(UriTemplate = "contact/{idStr}")]
@@ -196,11 +246,22 @@ namespace PartyService
         {
             try
             {
-                int id = Convert.ToInt32(idStr);
+                int id;
+
+                if (!int.TryParse(idStr, out id))
+                {
+                    throw (new PartyServiceException(
+                        string.Format("'{0}' is not a valid contact identifier", idStr),
+                        HttpStatusCode.BadRequest));
+                }
+
                 DataRow r = database.GetContact(id);
+
                 if (r == null)
                 {
-                    throw (new Exception(string.Format("Contact {0} not found", id)));
+                    throw (new PartyServiceException(
+                        string.Format("Contact {0} not found", id),
+                        HttpStatusCode.NotFound));
                 }
 
                 string type = Convert.ToString(r["type"]);
@@ -221,14 +282,383 @@ namespace PartyService
                     case "T":
                         return new Telephone(id, subType, telephoneNumber);
                     default:
-                        return null;
+                        throw (new PartyServiceException(
+                        string.Format("Unknown contact type '{0}' retrieved", type)));
                 }
             }
             catch (Exception ex)
             {
-                Contact c = new Contact();
-                c.Error = new Error(ex.Message);
-                return c;
+                return (Contact)ErrorObject(new Contact(), ex);
+            }
+        }
+
+        [WebInvoke(UriTemplate="address")]
+        [OperationContract]
+        public Address CreateAddress(Address address)
+        {
+            try
+            {
+                int newId = database.CreateAddress(address.Street, address.Town, address.County, address.PostCode);
+                return (Address)GetContact(Convert.ToString(newId));
+            }
+            catch (Exception ex)
+            {
+                return (Address)ErrorObject(address, ex);
+            }
+        }
+
+        [WebInvoke(UriTemplate = "address/{idStr}")]
+        [OperationContract]
+        public Address UpdateAddress(string idStr, Address address)
+        {
+            try
+            {
+                int id;
+
+                if (!int.TryParse(idStr, out id))
+                {
+                    throw (new PartyServiceException(
+                        string.Format("'{0}' is not a valid contact identifier", idStr),
+                        HttpStatusCode.BadRequest));
+                }
+
+                if (database.GetContact(id) == null)
+                {
+                    throw (new Exception(string.Format("Address {0} not found", id)));
+                }
+
+                database.UpdateAddress(id, address.Street, address.Town, address.County, address.PostCode);
+
+                return (Address)GetContact(idStr);
+            }
+            catch (Exception ex)
+            {
+                return (Address)ErrorObject(address, ex);
+            }
+        }
+
+        [WebInvoke(UriTemplate = "email")]
+        [OperationContract]
+        public Email CreateEmail(Email email)
+        {
+            try
+            {
+                int newId = database.CreateEmail(email.SubType, email.Address);
+                return (Email)GetContact(Convert.ToString(newId));
+            }
+            catch (Exception ex)
+            {
+                return (Email)ErrorObject(email, ex);
+            }
+        }
+
+        [WebInvoke(UriTemplate = "email/{idStr}")]
+        [OperationContract]
+        public Email UpdateEmail(string idStr, Email email)
+        {
+            try
+            {
+                int id;
+
+                if (!int.TryParse(idStr, out id))
+                {
+                    throw (new PartyServiceException(
+                        string.Format("'{0}' is not a valid contact identifier", idStr),
+                        HttpStatusCode.BadRequest));
+                }
+
+                if (database.GetContact(id) == null)
+                {
+                    throw (new Exception(string.Format("Email {0} not found", id)));
+                }
+
+                database.UpdateEmail(id, email.SubType, email.Address);
+
+                return (Email)GetContact(idStr);
+            }
+            catch (Exception ex)
+            {
+                return (Email)ErrorObject(email, ex);
+            }
+        }
+
+        [WebInvoke(UriTemplate = "telephone")]
+        [OperationContract]
+        public Telephone CreateTelephone(Telephone telephone)
+        {
+            try
+            {
+                int newId = database.CreateTelephone(telephone.SubType, telephone.Number);
+                return (Telephone)GetContact(Convert.ToString(newId));
+            }
+            catch (Exception ex)
+            {
+                return (Telephone)ErrorObject(telephone, ex);
+            }
+        }
+
+        [WebInvoke(UriTemplate = "telephone/{idStr}")]
+        [OperationContract]
+        public Telephone UpdateTelephone(string idStr, Telephone telephone)
+        {
+            try
+            {
+                int id;
+
+                if (!int.TryParse(idStr, out id))
+                {
+                    throw (new PartyServiceException(
+                        string.Format("'{0}' is not a valid contact identifier", idStr),
+                        HttpStatusCode.BadRequest));
+                }
+
+                if (database.GetContact(id) == null)
+                {
+                    throw (new Exception(string.Format("Telephone {0} not found", id)));
+                }
+
+                database.UpdateTelephone(id, telephone.SubType, telephone.Number);
+
+                return (Telephone)GetContact(idStr);
+            }
+            catch (Exception ex)
+            {
+                return (Telephone)ErrorObject(telephone, ex);
+            }
+        }
+
+
+        [OperationContract]
+        [WebInvoke(UriTemplate = "contact/{idStr}/delete")]
+        public Acknowledgement DeleteContact(string idStr)
+        {
+
+            try
+            {
+                int id;
+
+                if (!int.TryParse(idStr, out id))
+                {
+                    throw (new PartyServiceException(
+                        string.Format("'{0}' is not a valid contact identifier", idStr),
+                        HttpStatusCode.BadRequest));
+                }
+
+                if (database.GetContact(id) == null)
+                {
+                    throw (new Exception(string.Format("Contact {0} not found", id)));
+                }
+
+                database.DeleteContact(id);
+
+                return new Acknowledgement(
+                    string.Format("Contact {0} deleted", id));
+            }
+            catch (Exception ex)
+            {
+                return (Acknowledgement)ErrorObject(new Acknowledgement(), ex);
+            }
+
+        }
+
+        [WebGet(UriTemplate = "/party/{partyIdStr}/contact/{contactIdStr}")]
+        [OperationContract]
+        PartyContact GetPartyContact(string partyIdStr, string contactIdStr)
+        {
+
+            try
+            {
+
+                int partyId, contactId;
+
+                if (!int.TryParse(partyIdStr, out partyId))
+                {
+                    throw (new PartyServiceException(
+                        string.Format("'{0}' is not a valid party identifier", partyIdStr),
+                        HttpStatusCode.BadRequest));
+                }
+
+                if (!int.TryParse(contactIdStr, out contactId))
+                {
+                    throw (new PartyServiceException(
+                        string.Format("'{0}' is not a valid contact identifier", contactIdStr),
+                        HttpStatusCode.BadRequest));
+                }
+
+                DataRow r = database.GetPartyContact(partyId, contactId);
+
+                if (r == null)
+                {
+                    throw (new PartyServiceException(
+                        string.Format("Party {0} is not linked to contact {1}", partyId, contactId),
+                        HttpStatusCode.NotFound));
+                }
+
+                return new PartyContact(
+                    Convert.ToInt32(r["party_id"]),
+                    Convert.ToInt32(r["contact_id"]),
+                    Convert.ToDateTime(r["valid_from"]),
+                    Convert.ToDateTime(r["valid_until"]));
+            }
+            catch (Exception ex)
+            {
+                return (PartyContact)ErrorObject(new PartyContact(), ex);
+            }
+     
+        }
+
+
+        [WebInvoke(UriTemplate = "/party/contact")]
+        [OperationContract]
+        PartyContact CreatePartyContact(PartyContact partyContact)
+        {
+
+            try
+            {
+                database.CreatePartyContact(
+                    partyContact.PartyId, partyContact.ContactId, 
+                    partyContact.ValidFrom, partyContact.ValidUntil);
+
+                return GetPartyContact(
+                    Convert.ToString(partyContact.PartyId),
+                    Convert.ToString(partyContact.ContactId));
+
+            }
+            catch (Exception ex)
+            {
+                return (PartyContact)ErrorObject(partyContact, ex);
+            }
+
+        }
+
+        [WebInvoke(UriTemplate = "/party/{partyIdStr}/contact/{contactIdStr}")]
+        [OperationContract]
+        PartyContact UpdatePartyContact(string partyIdStr, string contactIdStr, PartyContact partyContact)
+        {
+
+            try
+            {
+
+                int partyId, contactId;
+
+                if (!int.TryParse(partyIdStr, out partyId))
+                {
+                    throw (new PartyServiceException(
+                        string.Format("'{0}' is not a valid party identifier", partyIdStr),
+                        HttpStatusCode.BadRequest));
+                }
+
+                if (!int.TryParse(contactIdStr, out contactId))
+                {
+                    throw (new PartyServiceException(
+                        string.Format("'{0}' is not a valid contact identifier", contactIdStr),
+                        HttpStatusCode.BadRequest));
+                }
+
+                DataRow r = database.GetPartyContact(partyId, contactId);
+
+                if (r == null)
+                {
+                    throw (new PartyServiceException(
+                        string.Format("Party {0} is not linked to contact {1}", partyId, contactId),
+                        HttpStatusCode.NotFound));
+                }
+
+                database.UpdatePartyContact(partyId, contactId, partyContact.ValidFrom, partyContact.ValidUntil);
+
+                return GetPartyContact(partyIdStr, contactIdStr);
+            }
+            catch (Exception ex)
+            {
+                return (PartyContact)ErrorObject(partyContact, ex);
+            }
+
+        }
+
+        [WebInvoke(UriTemplate = "/party/{partyIdStr}/contact/{contactIdStr}/delete")]
+        [OperationContract]
+        Acknowledgement DeletePartyContact(string partyIdStr, string contactIdStr)
+        {
+
+            try
+            {
+
+                int partyId, contactId;
+
+                if (!int.TryParse(partyIdStr, out partyId))
+                {
+                    throw (new PartyServiceException(
+                        string.Format("'{0}' is not a valid party identifier", partyIdStr),
+                        HttpStatusCode.BadRequest));
+                }
+
+                if (!int.TryParse(contactIdStr, out contactId))
+                {
+                    throw (new PartyServiceException(
+                        string.Format("'{0}' is not a valid contact identifier", contactIdStr),
+                        HttpStatusCode.BadRequest));
+                }
+
+                DataRow r = database.GetPartyContact(partyId, contactId);
+
+                if (r == null)
+                {
+                    throw (new PartyServiceException(
+                        string.Format("Party {0} is not linked to contact {1}", partyId, contactId),
+                        HttpStatusCode.NotFound));
+                }
+
+                database.DeletePartyContact(partyId, contactId);
+
+                return new Acknowledgement(
+                    string.Format("Contact {0} removed from party {1}", contactId, partyId));
+            }
+            catch (Exception ex)
+            {
+                return (Acknowledgement)ErrorObject(new Acknowledgement(), ex);
+            }
+
+        }
+
+        [WebGet(UriTemplate="/party/{idStr}/contacts")]
+        [OperationContract]
+        public ContactResults GetPartyContacts(string idStr)
+        { 
+            try
+            {
+
+                int id;
+                string baseUrl = WebOperationContext.Current.IncomingRequest.UriTemplateMatch.BaseUri.AbsoluteUri;
+
+                if (!int.TryParse(idStr, out id))
+                {
+                    throw (new PartyServiceException(
+                        string.Format("'{0}' is not a valid party identifier", idStr),
+                        HttpStatusCode.BadRequest));
+                }
+
+                ContactResults contactResults = new ContactResults();
+
+                DataTable t = database.GetContactByPartyId(id);
+
+                foreach (DataRow r in t.Rows)
+                {
+                    int contactId = Convert.ToInt32(r["id"]);
+                    string type = Convert.ToString(r["sub_type"]);
+                    string detail = Convert.ToString(r["detail"]);
+                    string link = string.Format("{0}contacts/{1}", 
+                        baseUrl, contactId);
+
+                    contactResults.ContactList.Add
+                        (new ContactSummary(id, type, detail,link));
+                }
+
+                return contactResults;
+                        
+            }
+            catch (Exception ex)
+            {
+                return (ContactResults)ErrorObject(new ContactResults(), ex);
             }
         }
